requirements.txt
flask==2.0.1
pillow==10.0.0
tensorflow==2.15.0
numpy==1.23.5
werkzeug==2.0.1 